requirements.txt
requests==2.31.0
python-dotenv==1.0.0
outscraper
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
pytest==7.4.3